requirements.txt
about-time==4.2.1
aiofiles==23.2.1
aiogram==3.9.0
aiohappyeyeballs==2.3.5
aiohttp==3.9.5
aiohttp-socks==0.7.0
aiosignal==1.3.1
alive-progress==3.1.5
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
autoslot==2022.12.1
beautifulsoup4==4.12.3
beautifultable==1.1.0
cdp-socket==1.2.8
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
exceptiongroup==1.2.2
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
grapheme==0.6.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
ImageHash==4.3.1
magic-filter==1.0.12
multidict==6.0.5
numpy==2.0.1
orjson==3.10.7
outcome==1.3.0.post0
pillow==10.4.0
platformdirs==4.2.2
protobuf==5.27.3
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-socks==2.5.0
PyWavelets==1.6.0
requests==2.32.3
scipy==1.14.0
selenium==4.23.1
selenium_driverless==1.9.3.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
